Move dependences from cocoapods to spm
