+++ Packages/Utilities/Package.swift
@@ -14,14 +14,14 @@ let package = Package(
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
+         .package(url: "git@github.com:aalzres/SwifterSwift.git", from: "5.2.0")
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "Utilities",
-            dependencies: []),
+            dependencies: ["SwifterSwift"]),
         .testTarget(
             name: "UtilitiesTests",
             dependencies: ["Utilities"]),
