+++ Packages/Utilities/Sources/Utilities/Extension/SwiftStdlib/String+HashMD5.swift
@@ -1,29 +1,12 @@
 //
-//  String+Extension.swift
-//  PORTFOLIO
+//  File.swift
+//  
 //
-//  Created by Andres Felipe Alzate on 16/06/2020.
-//  Copyright © 2020 aalzres. All rights reserved.
+//  Created by Andres Felipe Alzate Restrepo on 10/12/21.
 //
 
 import Foundation
 
-extension String {
-    var isValidEmail: Bool {
-        let emailRegEx = "[A-Z0-9a-z._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,64}"
-        
-        let emailPred = NSPredicate(format:"SELF MATCHES %@", emailRegEx)
-        return emailPred.evaluate(with: self)
-    }
-    
-    func localized(bundle: Bundle = .main, tableName: String = "Localizable") -> String {
-        return NSLocalizedString(self, tableName: tableName, value: "**\(self)**", comment: "")
-    }
-    
-    func localizedWithFormat(arguments: CVarArg...) -> String{
-       return String(format: self.localized(), arguments: arguments)
-    }
-}
 // MARK: - Hash md5
 import var CommonCrypto.CC_MD5_DIGEST_LENGTH
 import func CommonCrypto.CC_MD5
@@ -34,7 +17,7 @@ extension String {
         let length = Int(CC_MD5_DIGEST_LENGTH)
         let messageData = self.data(using:.utf8)!
         var digestData = Data(count: length)
-        
+
         _ = digestData.withUnsafeMutableBytes { digestBytes -> UInt8 in
             messageData.withUnsafeBytes { messageBytes -> UInt8 in
                 if let messageBytesBaseAddress = messageBytes.baseAddress, let digestBytesBlindMemory = digestBytes.bindMemory(to: UInt8.self).baseAddress {
@@ -46,8 +29,6 @@ extension String {
         }
         return digestData
     }
-    
-    func hashMD5() -> String {
-        return self.md5().map { String(format: "%02hhx", $0) }.joined()
-    }
+
+    public func hashMD5() -> String { md5().map { String(format: "%02hhx", $0) }.joined() }
 }
